requirements.txt
# WattaBuzz 프로젝트 종속성
# 검증된 버전들 (새 설치 환경에서 테스트됨)

# Web Framework
streamlit==1.47.1

# Data Collection & APIs
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
requests==2.32.4

# Data Processing & ML
pandas==2.3.1
numpy==2.0.2
transformers==4.54.0
scikit-learn==1.6.1
torch==2.7.1
tokenizers==0.21.2
safetensors==0.5.3

# Data Visualization
plotly==5.24.1
matplotlib==3.9.4
contourpy==1.2.1

# Database
psycopg2-binary==2.9.10
sqlalchemy==1.4.53

# Utilities
python-dotenv==1.1.1
schedule==1.2.2
pydantic==2.5.3

# Airflow (필수 - DAG 워크플로우용)
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.8.0

# Flask 관련 (Airflow 종속성)
Flask==2.2.5
Flask-Session==0.5.0
Flask-AppBuilder==4.3.10
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1

# 기타 필수 종속성
PyYAML==6.0.2
click==8.1.8
Jinja2==3.1.6

# macOS 호환성
urllib3<2.0

# Airflow API 문서 (선택사항)
# connexion[swagger-ui] 